Initial scripting support
This change adds a PythonOperation class, which performs an operation
that is defined by a Python script using Jython.  The idea behind this
is to allow users to extend the program with new, more complex
operations using an easy scripting language, rather than requiring them
to modify the program itself.  This also makes it possible to share
"plugins" designed for a specific task that may be useful to other
people.