Add Sink
With this change, sockets can be set to be "published", and classes
implementing the Sink interface listen for published values.  An example
of this is the NetworkTables backend, which will be implemented as a
Sink.

For now, there is a DummySink class that just prints any published
values to the console.