Add a new window for displaying published sockets
We should have some way of simply viewing all published socket values
regardless of the current sink, so the user doesn't have to load up a
NetworkTables client or inspect the console output to see if the right
values are showing up.

This commit adds a window that shows a table of published sockets.
Their values are updated as soon as they change by having the controller
subscribe to SocketPublishedEvents on the event bus.

The FXML, CSS, and Java controller are included, but this commit does
not add the driver code to create the window.  We should do that once
more progress has been made on the GUI, but for now I have a separate
stash with a modified Main.java to showcase this feature.