Make pipeline react to step added & removed events
Instead of just taking the initial list of steps, Pipeline should listen
for new steps being added and removed.  The also made it necessary to
add getters for the underlying core model for StepView, so I also added
similar getters to pipelineview and socketcontrolview, since those will
also be needed.