Add ConnectionView and handlers for adding/removing connections
PipelineView basically listens for ConnectionAddedEvents and
ConnectionRemovedEvents and adds and removes the ConnectionView
controls as needed.  The only difference between this and most of the
similar setups (Steps and Sockets) is that the PipelineView must look up
the location of the "handles", which are the points on each of the
sockets that the connections connect to.

This also makes a couple changes to the main classes to include some
sample driver code.  As more features are implemented, this code will
change to showcase them without having some temporary stash somewhere
with the throwaway driver code.  As the application becomes more
complete, we can remove more and more of this code and have the actual
functionality to kick off the features being demoed in places.