Keep track of all connections in the socket class
Sockets really need to know what connections they have instead of just a
flag saying weather or not there is a connection.  This is necessary for
several of the gui operations, and also may be used in the future to
make operations smarter about reusing buffers from step to step if they
only flow linearly through a series of steps.