Allow the user to add connections
This is just the initial, very basic support for adding connections.
The user can click a socket, and then click another socket to connect
them.  The style of the sockets is changed based on weather they're
connected, not connected, or in the processes of being connected (but
the second socket hasn't been selected yet).

TODO:
    - Show a curve similar to a ConnectionView while the user is
      creating a connection
    - Add some way to delete connections.  Performing some action on a
      socket isn't good enough (since one socket can have multiple
      connections), so it might be that the user clicks the connection
      itself, or only the input socket.
    - Fix a bug where connections don't reposition themselves when the
      sockets move.  This happends when a step is deleted, causing the
      connections from all of the steps that shift over to just stay
      floating there.