Add SocketHint default values with Suppliers
Instead of specifying a default value for sockets where no other value
is set, you can specify a function that returns an initial value for any
socket that isn't initialized with a value.  This ensures that all
sockets have a value, and also that those values can each be a new
instance (for mutable types like Mat).

For example, to make every socket have 5 as an initial value:

    sh = new SocketHint<>("my-socket", Integer.class, 5);

But to make every socket construct a new Mat as an initial value (so
they don't all share the same instance):

    sh = new SocketHint<>("my-socket", Mat.class, Mat::new);