Add a "publish" toggle button on output sockets
This button allows the user to specify which sockets are published to
the current sink.

Instead of adding in more `if (isInputSocket())` statements to
SocketControlView, the class has been replaced by two separate classes
for input sockets and output sockets.  The only thing that the views
have in common is that they both have a label and a SocketHandleView,
and those aren't even in the same place, so it doesn't make sense to
have them as the same class or as subclasses of a common class.