Add a preview pane
When sockets are previewed, the preview pane looks up the right type of
view to use for that socket and creates and automatically-updating
real-time preview for it.

Right now, the only implementation simply renders to result of callint
`toString()` on the socket's value (which is good for things like
numbers), but more implementations will need to be written for things
like previewing images.