Don't try to be clever about not sending SocketChangedEvents
In Socket::setValue(), it doesn't really make sense to not send a
SocketChangedEvent if the new value is equal to the old value, since
some types (like Mat) can actually be modified in-place.

Since the "optimization" of not sending events doesn't really come up in
practice anywhere intentional, it's better just to remove this check.

For this to not break everything, we have to make sure that steps only
update if one of their inputs changes, not if *any* of their sockets
changes.  It should probably be like this anyways, because otherwise
we're running twice as many operations as we need to.