Move the gompei samples from a python script into a samples directory
The gompei.py operation was a temporary workaround until we had image
sources working natively.  We should still keep the images around for
easy testing, so I made a "samples" directory in the project.

Once opening Python scripts also works, it'll probably be a good idea to
move throwaway or demo scripts there as well (although most scripts will
be included in the application resources).