Make Sources have names
The Source interface now has a getName() method, which returns some kind
of identifier or filename.  One consequence of this is that we can't
easily reuse the same Source object with a different file/webcam/etc,
but we don't do this anyways, so it doesn't affect much if we just only
load the source in the constructor.

Now the GUI can actually have a descriptive label on each source instead
of just the class name.