Only make sockets publishable if their socket hint says so
This adds a "publishable" field to SocketHint that controls if a socket
can be set to be published to a sink.  It doesn't make sense for a lot
of sockets to be able to be published

Not preventing it would create confusion (mats would likely be published
somewhere as the result of Mat::toString(), which is almost certainly
not was the user things).  Plus, the publish buttons create visual
clutter.

Since the default setting is `false`, we should make sure that anything
that could be published is set a publishable.