Refactor most InputSocketView types into subclasses
Instead of having a huge switch statement with a bunch of checkArguments
and other conditionals, it's much easier to maintain a different
subclass for each type of input.  Right now, CheckboxInputSocketView,
SliderInputSocketView, and SpinnerInputSocketView are implemented.

With the logic of choosing what view to use encapsulated into a factory,
it should be easier to add more advanced views without adding stuff to a
large, multi-purpose InputSocketView class.