Update image previews in the UI thread
This is probably the cause of the mysterious scenegraph
NullPointerException that came up before.  The scenegraph should only be
manipulated from the UI thread, and the preview code is actually running
in the thread that eventBus.post was called in, which is the webcam
executor thread.  When multiple threads are touchin the scenegraph
probably hundreds of times per second, a race condition occasionally
happens.

We'll probably need some stress testing to make sure it's actually
fixed.