Prevent sockets from keeping old values after connections are removed
First of all, input sockets reset themselves to their default value when
they are no longer connected to anything.  This, for example, makes sure
that the input to a step that takes a Mat doesn't keep a reference to
a mat that keeps getting updated with new camera data after it's been
disconnected from the camera source.

In addition, operations that short circuit when given default values
(such as most operations that take non-empty Mats) should also set their
outputs to the default values.  This could be done automatically once we
have a way to specify in SocketHint which values are required instead of
doing the same check inside of each operation.

Fixes #47