Create StepControllers with Guice
StepControllers are different from the stuff ported over so far since
they depend on a non-injection parameter (the step).

Guice solves this by letting you bind a factory interface for
StepControllers.  The framework automatially creates an implementation
of the interface that still does injection magic, so we can just inject
the factory and call the create method with whatever parameters we need.

One downside to this is that constructing new objects with such
dependencies is a few more lines of code to set the controller on an
FXMLLoader (it was just new StepView() before).  However, this code just
used to be placed in the constructor for the view instead of being
written out by the caller.

The extra verbosity is also more of a consequence of separating views
and controllers to not be the same class, which is definitely a good
thing even with this slightly annoying corner case.