Fix Platform.runLater() usage
Our code is not really written to handle asyncronous GUI updates in many
situations, such as updating the pipeline view when steps are
added/deleted/moved.  So, in these situations we should block the
calling thread until the GUI thread has finished performing the
specified action.

In addition, it is not necessary to call Platorm.runLater() from the GUI
thread.