Fixes ImageConverter running in the wrong thread
If an image changed dimention in the preview pane
it's UI counterpart could be changed in the wrong thread.
This could cause random ArrayOutOfBounds Exceptions without
any context as to what was causing the problem.
This forces the ImageConverter.convert method to be run in the
UI thread or it will throw an exception.

Possibly related: #225