Exception alert enhancements
GitHub's servers (and many modern browsers) apparently don't like us
redirecting users to thousand-character-long URLs to automatically fill
in issue forms.  So, I got rid of the whole thing and just put the
generated text in the exception alert.  Users can copy and paste it
themselves.

The list of system properties was also updated to remove some
redundant/useless ones.  We can always ask for more info in necessary,
but the ones removes can all be derived from existing ones.  This should
just clear up some visual clutter.

The lack of styling in the alert has also bothered me for a while, so
that's fixed now.