Fix windows hanging when being shutdown
This was caused because a shtdown hook threw an exception.
Our main application is designed to try to handle threads dying by
displaying a UI element. This should not be done in a shutdown hook.
What we have to do is ensure that when we try dto display the
exception alert it isn't from inside a shutdown hook.

Closes #297