Add a "Publish Video" operation
This allows you to view the video feed with any compatible FRC dashboard
(tested with SmartDashboard, but the FRC dashboard uses the same
protocol when it's in "hardware" mode).  Intermediate steps can also be
published, allowing teams to visualize parts of the pipeline even when
it's running in headless mode.