Add deploy shortcut (ctrl+r) and update icon
The dialog graphic is now the settings icon, since it shows deploy
settings.  It also looked weird before, since it had the same play icon
in both a button and the dialog header. Plus, I'm somewhat concerned
that the unlabled play icon could be interpreted as a status icon (ie:
it's indicating that something is currently running)