Code Generation (#597)
Add support for generating Java, C++, and Python code from a pipeline.

Exported code uses the official OpenCV API for that language (so no JavaCV). The generated code is set up so that a user supplies an image, runs the pipeline, and gets the data outputs manually. There is no integration with image sources, such as webcams, to make the generated code as portable as possible


Notable changes and additions:

* Made file chooser default to Pipeline as name and made exporter name the class properly

* Created an icon for code generation

* Adds new switches for gradle:
-Pgeneration to compile/run generation and non generation tests
-Pgenonly to compile all tests but only run generation tests
if neither is specified then generation tests will not be compiled
and thus OpenCV is not required to compile tests other than generation

* Always display the correct language if tests fail.

* Added shortcut for code generation

* Updated to use a new thread for export

* Made code generation work from within an executable.

Also changed the template to the code generation folder.
This made it easier for templates to be found.

* Don't try generating code for publishing operations

* Make variable and method names fit standard language conventions.

Variables and methods no longer have numbers if there's only one of that operation, or if an operation only has one output.
Operations with multiple outputs have the outputs named the same as on the socket (instead of just 'output0', 'output1', ...)
Rename `MutableOf` to `Ref` because it's really just holding a reference and it's used similarly to output parameters in C/C++

* Add alert when trying to generate code for a pipeline with non-exportable operations

* Fixed Java and Python tests on Travis. C++ test are disabled.

* Use OpenCV artifacts on FRC maven for code gen testing. Enable code gen tests on Travis' OSX platform