Fixes #920 incorrect java codegen for filter lines
Original VM template was reassigning the filtered list to the outputs variable
which would have no effect on the contents of the passed-in argument.  Working
pipelines in the GRIP GUI would show zero filtered lines when the generated class
was used.

Altered the template to manipulate the passed-in list object instead, using
`clear()` and `addAll()`.

Note: there doesn't appear to be unit tests for the generated classes, and I
don't have time to write them.  I tested the fix using test cases in our
own FRC project that use the generated class.